files
STL files for 3D printing a case for the YARD Stick One and Ubertooth usb devices from Great Scott Gadgets. These Files are from Miek ='s Gist Github, I am only posting them so that they are a little easier to find. I had problems finding these and I actually went thru the whole process of learning Fusion 360 and making my own prior to finding these. However the ones that I designed myself are nowhere as nice looking as these ones. So with that said I take no credit in the designing of these or the device itself. Thank You to Mike O. for designing such a wonderfully  easy to use device and Thank You to Miek for posting these on his Gist.